Add Eliza.lua as resume fodder
I saw in the Freedos release on christmas that they included Eliza in their
distribution, and I had the weird 10pm drive to reimplement her in Lua. This
turned into a fun trip of studying her source in Lisp, Pascal, and Go to try
and deduce what the trick was. Turned out it wasn't very complex, it only
looked complex because the technical limitations from the earliest versions
kept getting ported forward. This version should reveal how very simple her
mechanisms are.